+++ packages/search-select/src/search-select.tsx
@@ -36,6 +36,7 @@ import {
   shallowRef,
   watch,
 } from 'vue';
+import { type SlotsType } from 'vue';
 
 import { useLocale, usePrefix } from '@bkui-vue/config-provider';
 import { clickoutside } from '@bkui-vue/directives';
@@ -45,6 +46,7 @@ import { debounce } from '@bkui-vue/shared';
 import SearchSelectInput from './input';
 import SearchSelected from './selected';
 import {
+  // DeleteBehavior,
   GetMenuListFunc,
   ICommonItem,
   ISearchItem,
@@ -102,6 +104,13 @@ export const SearchSelectProps = {
       return [ValueBehavior.ALL, ValueBehavior.NEEDKEY].includes(v);
     },
   },
+  // deleteBehavior: {
+  //   type: String as PropType<`${DeleteBehavior}`>,
+  //   default: DeleteBehavior.CHAR,
+  //   validator(v: DeleteBehavior) {
+  //     return [DeleteBehavior.CHAR, DeleteBehavior.VALUE].includes(v);
+  //   },
+  // },
 };
 export default defineComponent({
   name: 'SearchSelect',
@@ -109,7 +118,13 @@ export default defineComponent({
     clickoutside,
   },
   props: SearchSelectProps,
-  emits: ['update:modelValue', 'search'],
+  emits: ['update:modelValue', 'search', 'selectKey'],
+  slots: Object as SlotsType<{
+    menu: MenuSlotParams;
+    prepend: void;
+    append: void;
+    validate: void;
+  }>,
   setup(props, { emit }) {
     const t = useLocale('searchSelect');
     const { resolveClassName } = usePrefix();
@@ -124,7 +139,7 @@ export default defineComponent({
     });
 
     // refs
-    const inputRef = ref<typeof SearchSelectInput>(null);
+    const inputRef = ref<InstanceType<typeof SearchSelectInput>>(null);
     const wrapRef = ref<HTMLDivElement>(null);
 
     // vars
@@ -214,6 +229,8 @@ export default defineComponent({
       onEditBlur,
       onValidate,
       editKey,
+      searchData: computed(() => props.data),
+      isClickOutside: handleInputOutside,
     });
     function onEditClick(item: SelectedItem, index: number) {
       editKey.value = `${item.id}_${index}`;
@@ -266,12 +283,13 @@ export default defineComponent({
     }
     function handleWrapClick() {
       if (!editKey.value) {
-        inputRef.value.handleInputFocus();
+        inputRef.value.inputFocusForWrapper();
       }
     }
     function handleClearAll() {
       selectedList.value = [];
       overflowIndex.value = -1;
+      inputRef.value.inputClearForWrapper();
       emit('update:modelValue', []);
     }
     function handleInputOutside(target: Node) {
@@ -303,7 +321,11 @@ export default defineComponent({
       isFocus.value = v;
     }
     function handleClickSearch(e: MouseEvent) {
+      inputRef.value.inputEnterForWrapper();
       emit('search', e);
+    }
+    function handleSelectedKey(a: any) {
+      emit('selectKey', a);
     }
     return {
       inputRef,
@@ -325,6 +347,7 @@ export default defineComponent({
       handleClickSearch,
       localConditions,
       resolveClassName,
+      handleSelectedKey,
       t,
     };
   },
@@ -366,6 +389,7 @@ export default defineComponent({
               validateValues={this.validateValues}
               valueBehavior={this.valueBehavior as ValueBehavior}
               onDelete={this.handleDeleteSelected}
+              onSelectKey={this.handleSelectedKey}
               v-slots={{ ...menuSlots }}
             />
             <div class='search-container-input'>
@@ -383,6 +407,7 @@ export default defineComponent({
                 onAdd={this.handleAddSelected}
                 onDelete={this.handleDeleteSelected}
                 onFocus={this.handleInputFocus}
+                onSelectKey={this.handleSelectedKey}
                 v-slots={{ ...menuSlots }}
               />
             </div>