+++ packages/search-select/src/selected.tsx
@@ -62,9 +62,10 @@ export default defineComponent({
     validateValues: Function as PropType<ValidateValuesFunc>,
     valueBehavior: String as PropType<ValueBehavior>,
   },
-  emits: ['delete'],
+  emits: ['delete', 'selectKey'],
   setup(_props, { emit }) {
-    const inputRef = ref<typeof SearchSelectInput>(null);
+    const inputRef = ref<InstanceType<typeof SearchSelectInput>>(null);
+    const selectedInputRef = ref<HTMLDivElement>(null);
     const { onEditClick, onEditEnter, onEditBlur, editKey } = useSearchSelectInject();
     function handleDeleteSelected(index: number) {
       emit('delete', index);
@@ -73,6 +74,11 @@ export default defineComponent({
       e.preventDefault();
       e.stopPropagation();
       onEditClick(item, index);
+      emit('selectKey', {
+        id: item.id,
+        name: item.name,
+        values: item.values.slice(),
+      });
       // magic code
       setTimeout(() => inputRef.value.handleInputFocus(), 200);
     }
@@ -83,8 +89,8 @@ export default defineComponent({
       if (isFocus) return;
       onEditBlur();
     }
-    function handleInputOutside() {
-      return true;
+    function handleInputOutside(target: Node) {
+      return !selectedInputRef.value?.contains(target);
     }
     function copySeletedItem(item: SelectedItem): SelectedItem {
       const newItem = new SelectedItem(item.searchItem, item.type);
@@ -94,6 +100,7 @@ export default defineComponent({
     }
     return {
       inputRef,
+      selectedInputRef,
       editKey,
       copySeletedItem,
       handleDeleteSelected,
@@ -109,6 +116,7 @@ export default defineComponent({
         <div
           class='selected-input'
           key={this.editKey.toString()}
+          ref='selectedInputRef'
         >
           <SearchSelectInput
             ref='inputRef'
@@ -124,6 +132,7 @@ export default defineComponent({
             valueBehavior={this.valueBehavior}
             onAdd={v => this.handleAddSelected(v, index)}
             onFocus={this.handleInputFocus}
+            v-slots={{ ...this.$slots }}
           />
         </div>
       ) : (