+++ packages/table/src/use-render.tsx
@@ -42,11 +42,11 @@ import Settings from './plugins/settings';
 import useDraggable from './plugins/use-draggable';
 import useFixedColumn from './plugins/use-fixed-column';
 import useHeadCell from './plugins/use-head-cell';
+import useShiftKey from './plugins/use-shift-key';
 import { Column, Settings as ISettings, TablePropTypes } from './props';
 import { ITableResponse } from './use-attributes';
 import {
   formatPropAsArray,
-  getRowKeyNull,
   getRowText,
   isRowSelectEnable,
   resolveCellSpan,
@@ -61,7 +61,6 @@ export default (
   props: TablePropTypes,
   context: SetupContext<any>,
   tableResp: ITableResponse,
-  styleRef,
   head,
   root,
   resetTableHeight,
@@ -75,6 +74,9 @@ export default (
   const columns = computed(() => formatData.value.columns);
 
   const settings = computed(() => formatData.value.settings);
+
+  const { isShiftKeyDown, getStore, setStore } = useShiftKey(props);
+
   // const activeSortIndex = ref(null);
 
   /**
@@ -193,8 +195,15 @@ export default (
   };
 
   const handlePageChange = (current: number) => {
-    Object.assign(props.pagination, { current, value: current });
-    context.emit(EMIT_EVENTS.PAGE_VALUE_CHANGE, current);
+    if (typeof props.pagination === 'object' && current !== props.pagination.current) {
+      Object.assign(props.pagination, { current, value: current });
+      context.emit(EMIT_EVENTS.PAGE_VALUE_CHANGE, current);
+      return;
+    }
+
+    if (typeof props.pagination === 'boolean' && props.pagination !== false) {
+      context.emit(EMIT_EVENTS.PAGE_VALUE_CHANGE, current);
+    }
   };
 
   /**
@@ -209,7 +218,7 @@ export default (
       backgroundColor: props.thead.color,
     };
 
-    const { resolveFixedColumnStyle } = useFixedColumn(props, tableResp, head);
+    // const { resolveFixedColumnStyle } = useFixedColumn(props, tableResp, head);
 
     return (
       <>
@@ -218,7 +227,7 @@ export default (
             <tr>
               {filterColGroups.value.map((column, index: number) => {
                 const { getTH } = useHeadCell(props, context, column, tableResp);
-                const headStyle = Object.assign({}, resolveFixedColumnStyle(column, styleRef.value.hasScrollY), {
+                const headStyle = Object.assign({}, resolveFixedColumnStyle(column), {
                   '--background-color': DEF_COLOR[props.thead?.color ?? IHeadColor.DEF1],
                 });
 
@@ -310,6 +319,7 @@ export default (
       ...formatPropAsArray(props.rowStyle, [row, rowIndex]),
       {
         '--row-height': `${getRowHeight(row, rowIndex)}px`,
+        '--scroll-x': `${tableResp.formatData.layout.translateX}px`,
       },
     ];
 
@@ -319,13 +329,14 @@ export default (
       rowIndex % 2 === 1 && props.stripe ? 'stripe-row' : '',
     ];
     const rowId = tableResp.getRowAttribute(row, TABLE_ROW_ATTRIBUTE.ROW_UID);
+
     return [
       <TableRow key={rowId}>
         <tr
           // @ts-ignore
           style={rowStyle}
           class={rowClass}
-          key={getRowKeyNull(row, props, rowIndex)}
+          key={rowId}
           data-row-index={rowIndex}
           onClick={e => handleRowClick(e, row, rowIndex, rows)}
           onDblclick={e => handleRowDblClick(e, row, rowIndex, rows)}
@@ -380,16 +391,20 @@ export default (
                 context.emit(type, args);
               };
 
+              const columnKey = `${rowId}_${index}`;
+              const cellKey = `${rowId}_${index}_cell`;
               return (
                 <td
                   class={cellClass}
                   style={cellStyle}
                   colspan={colspan}
                   rowspan={rowspan}
+                  key={columnKey}
                   onClick={event => handleEmit(event, EMIT_EVENTS.CELL_CLICK)}
                   onDblclick={event => handleEmit(event, EMIT_EVENTS.CELL_DBL_CLICK)}
                 >
                   <TableCell
+                    key={cellKey}
                     class={tdCtxClass}
                     column={column}
                     row={row}
@@ -508,6 +523,34 @@ export default (
       context.emit(EMIT_EVENTS.ROW_SELECT_CHANGE, { row, index, checked: value, data: props.data });
     };
 
+    const beforeRowChange = () => {
+      if (isShiftKeyDown.value && !isAll) {
+        const result = setStore(row, index);
+        if (result) {
+          const { start, end } = getStore();
+          const startIndex = start.index < end.index ? start.index : end.index;
+          const endIndex = start.index < end.index ? end.index : start.index;
+
+          (tableResp.pageData.slice(startIndex, endIndex + 1) ?? []).forEach(item => {
+            tableResp.setRowSelection(item, true);
+          });
+        }
+
+        context.emit(EMIT_EVENTS.ROW_SELECT, { row, index, checked: true, data: props.data, isShiftKeyDown: true });
+        context.emit(EMIT_EVENTS.ROW_SELECT_CHANGE, {
+          row,
+          index,
+          checked: true,
+          data: props.data,
+          isShiftKeyDown: true,
+        });
+
+        return Promise.resolve(!result);
+      }
+
+      return Promise.resolve(true);
+    };
+
     const indeterminate = tableResp.getRowAttribute(row, TABLE_ROW_ATTRIBUTE.ROW_SELECTION_INDETERMINATE);
     const isChecked = tableResp.getRowAttribute(row, TABLE_ROW_ATTRIBUTE.ROW_SELECTION);
     const isEnable = isRowSelectEnable(props, { row, index, isCheckAll: isAll });
@@ -518,6 +561,7 @@ export default (
         disabled={!isEnable}
         modelValue={isChecked}
         indeterminate={indeterminate as boolean}
+        beforeChange={beforeRowChange}
       />
     );
   };