+++ site/views/search-select/index.tsx
@@ -32,6 +32,7 @@ import PropsBox from '../../components/props-box';
 import { IPropsTableItem } from '../../typings';
 
 import BaseDemo from './base-demo.vue';
+import ComprehensiveUseDemo from './comprehensive-use.vue';
 import MenuDemo from './menu-demo.vue';
 import PlaceholderDemo from './placeholder-demo.vue';
 import RemoteDemo from './remote-demo.vue';
@@ -102,6 +103,13 @@ const propsJson: IPropsTableItem[] = [
     optional: ['all', 'need-key'],
   },
   {
+    name: 'comprehensive-use',
+    type: 'String',
+    default: 'delete-value',
+    desc: '配置按下delete键时是否删除整个value 配置为 delete-char 则删除当前字符 delete-value 则删除当前光标所在的整个value字符',
+    optional: ['delete-char', 'delete-value'],
+  },
+  {
     name: 'placeholder',
     type: 'String',
     default: '请选择',
@@ -129,7 +137,12 @@ const slotsJson = [
     type: 'scoped slot',
     default: [],
     desc: 'menu面板子项插槽',
-    params: '{list, item, hoverId, multiple, getSearchNode}',
+    params: `{
+      value: ICommonItem;
+      id: string;
+      name: string;
+      onSubmit: (value: string) => void;
+    }`,
   },
 ];
 const eventJson = [
@@ -211,6 +224,12 @@ const dataJson = [
     default: 'false',
     desc: '是否显示逻辑符号选项列表 默认不显示 仅在多选时生效',
   },
+  {
+    name: 'isCustomMenu',
+    type: 'Boolean',
+    default: 'false',
+    desc: '是否配置了自定义子项面板 仅在配置了menu插槽时生效',
+  },
 ];
 const slotColumnMap = {
   name: '名称',
@@ -275,8 +294,8 @@ export default defineComponent({
           <PlaceholderDemo />
         </DemoBox>
         <DemoBox
-          title='配置 menu 插槽来自定义 menu 面板'
-          subtitle='通过配置子选项属性 placeholder 来做到针对每一个选项都有独自的placeholder'
+          title='自定义 menu 面板'
+          subtitle='配置 menu 插槽来自定义 menu 面板'
           desc='menu 插槽'
           componentName='search-select'
           demoName='menu-demo'
@@ -288,9 +307,19 @@ export default defineComponent({
           subtitle='改变配置 valueBehevior 值为 need-key  来做到存文本不可生成 value tag'
           desc='valueBehevior 行为'
           componentName='search-select'
-          demoName='menu-demo'
+          demoName='value-behavior'
         >
           <ValueBehaviorDemo />
+        </DemoBox>
+
+        <DemoBox
+          title='综合使用'
+          subtitle='综合使用示例'
+          desc='综合使用示例'
+          componentName='search-select'
+          demoName='comprehensive-use'
+        >
+          <ComprehensiveUseDemo />
         </DemoBox>
         <PropsBox propsData={propsJson} />
         <PropsBox